Updates for Windows implementation. This code passes all regression tests except websocket tests, which have not been run yet - I expect many bugs there because this code has the first draft of Windows async file IO for the web server in O2.

+++ src/mqtt.cpp
@@ -338,7 +338,7 @@ void O2_MQTTcomm::disc_handler(char *payload, int payload_len)
         return;
     }
     int version;
-    if (!(version = o2_parse_version(vers_num, end - vers_num))) {
+    if (!(version = o2_parse_version(vers_num, (int) (end - vers_num)))) {
         #ifndef NDEBUG
                 // PRINT FOR DEBUGGING ONLY: payload IS NOT TERMINATED AND UNSAFE:
                 printf("o2_mqtt_disc_handler could not parse payload "
